avance fin de semana

+++ proyecto/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: vaguilar <vaguilar@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/11 12:54:51 by vaguilar          #+#    #+#             */
-/*   Updated: 2022/01/15 10:28:25 by vaguilar         ###   ########.fr       */
+/*   Updated: 2022/01/23 12:36:45 by vaguilar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *s, int c, size_t len)
 {
 	unsigned char	*ptr;
-	unsigned int	count;
+	size_t			count;
 
 	ptr = s;
 	count = 0;
